+++ bin/index.js
@@ -47,7 +47,7 @@ Deploying server side code:
 
   deploy                  => deploy the project
     --force               => deploy without confirmation
-    --stage <stage>       => optionally sets the stage to deploy to, defaults to 'dev'
+    --env <environment>   => optionally sets the environment to deploy to, defaults to 'dev'
 `
 
 const cli = meow({
@@ -59,12 +59,12 @@ const cli = meow({
   ],
   default: {
     'force': false,
-    'stage': 'dev'
+    'env': 'dev'
   },
   string: [
     'out',
     'port',
-    'stage'
+    'env'
   ]
 })