+++ commands/serverless.js
@@ -9,7 +9,7 @@ const lib = require('../lib/serverless.js')
 module.exports = function (input, flags, logger, options) {
   const cwd = options.cwd
   const out = flags.out
-  const stage = flags.stage
+  const env = flags.env
 
   if (!out) {
     return Promise.reject(new Error('"--out" is mandatory'))
@@ -18,6 +18,6 @@ module.exports = function (input, flags, logger, options) {
   return lib.copyProject(cwd, out)
     .then(() => lib.applyTemplate(out)) // TODO: eventually unnecessary?
     .then(() => lib.copyWrapper(out))
-    .then(() => lib.copyConfiguration(out, stage))
-    .then(() => lib.registerFunctions(out, stage))
+    .then(() => lib.copyConfiguration(out, env))
+    .then(() => lib.registerFunctions(out, env))
 }