+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 ### Changed
 
 -   Example project names to be more realistic
+
+-   **BREAKING CHANGE**: `--stage` flag to `--env`, functionality has not changed.